Fix tests with message/bundle removal
And bung a load of the Arbitrary instances in their own module under test/
