+++ src/Clapi/Types/Base.hs
@@ -6,7 +6,7 @@
 module Clapi.Types.Base
   ( Tag, unTag, mkTag
   , Time(..), TimeStamped(..)
-  , Attributee
+  , Attributee(..)
   , InterpolationLimit(..), Interpolation(..)
   , InterpolationType(..), interpolationType
   ) where
@@ -35,7 +35,7 @@ data Time = Time Word64 Word32 deriving (Eq, Show, Ord, Bounded)
 
 newtype TimeStamped a = TimeStamped (Time, a) deriving (Show, Functor)
 
-type Attributee = Text
+newtype Attributee = Attributee {unAttributee :: Text} deriving (Show, Eq)
 
 data InterpolationLimit = ILUninterpolated | ILConstant | ILLinear | ILBezier
   deriving (Show, Eq, Ord, Enum, Bounded)
